unreviewed

+++ PostgreSQL.17/Tests/Linq/DateTimeFunctionsTests/Tests.Linq.DateTimeFunctionsTests.SubDateMillisecond(PostgreSQL.17).sql
@@ -1,7 +1,7 @@
 ﻿-- PostgreSQL.17 PostgreSQL.15 PostgreSQL
 
 SELECT
-	ROUND(EXTRACT(EPOCH FROM ((t."DateTimeValue" + 2023456789 * Interval '1 Millisecond')::timestamp - t."DateTimeValue"::timestamp)) * 1000)
+	(ROUND(EXTRACT(EPOCH FROM ((t."DateTimeValue" + 2023456789 * Interval '1 Millisecond')::timestamp - t."DateTimeValue"::timestamp)) * 1000))::Float
 FROM
 	"LinqDataTypes" t
